Move to rye, add testing for py312 (#109)
* move project over to rye and clean up

* update workflows

* update readthedocs.yaml, add coverage run for pytest

* fix script for test_cov in pyproject.toml

* add xdoctest to workflow

* fix doctests -> np.float64 returned

needs to be explicitly stated in numpy>=2

* pin python to 3.9, add dev docs

* rtd theme for docs

* add editorconfig, update code formatting and link

+++ src/iniabu/elements.py
@@ -20,7 +20,7 @@ class Elements:
         'num_lin'
         >>> element = ini.ele["Si"]
         >>> element.abu_solar
-        999700.0
+        np.float64(999700.0)
 
     .. warning:: This class should NOT be manually created by the user. It is
         designed to be initialized by :class:`iniabu.IniAbu`.
